feat: 댓글, 상세페이지 css

+++ studywithme/src/components/CommentList.js
@@ -2,6 +2,7 @@ import React, { useEffect } from "react";
 import styled from "styled-components";
 import { useDispatch, useSelector } from "react-redux";
 import { actionCreators as commentActions } from "../redux/modules/comment";
+import { TiDeleteOutline } from "react-icons/ti";
 
 import Image from "../elements/Image";
 import Text from "../elements/Text";
@@ -12,9 +13,10 @@ const CommentList = (props) => {
   const dispatch = useDispatch();
   const comment = useSelector((state) => state.comment.list);
   const userId = useSelector((state) => state.user.user?.userId);
-
   const deleteComment = (commentId) => {
-    dispatch(commentActions.deleteCommentMiddleware(postId, commentId));
+    if (window.confirm("삭제하시겠습니까?") === true) {
+      dispatch(commentActions.deleteCommentMiddleware(postId, commentId));
+    }
   };
 
   useEffect(() => {
@@ -29,17 +31,20 @@ const CommentList = (props) => {
             <FlexGrid
               key={c.commentId}
               justify="space-between"
+              align="center"
               margin="5px 0px"
             >
               <FlexGrid align="center">
-                <Image size="30" />
-                <FlexGrid direction="column" margin="0px 10px">
-                  <Text size="15px">{c.userNickname}</Text>
-                  <Text size="10px" color="#cccccc">
-                    {time(c.date)}
-                  </Text>
+                <FlexGrid align="center" minWidth="120px">
+                  <Image size="36" />
+                  <FlexGrid direction="column" margin="0px 5px">
+                    <Text size="15px">{c.userNickname}</Text>
+                    <Text size="10px" color="#cccccc">
+                      {time(c.date)}
+                    </Text>
+                  </FlexGrid>
                 </FlexGrid>
-                <FlexGrid direction="column">
+                <FlexGrid direction="column" maxWidth="550px">
                   <Text>{c.textContent}</Text>
                   <FlexGrid>
                     <Button padding="0px">좋아요</Button>
@@ -47,14 +52,12 @@ const CommentList = (props) => {
                 </FlexGrid>
               </FlexGrid>
               {c.userId === userId ? (
-                <Button
-                  padding="0px"
-                  _onClick={() => {
+                <TiDeleteOutline
+                  className="deleteButton"
+                  onClick={() => {
                     deleteComment(c.commentId);
                   }}
-                >
-                  삭제
-                </Button>
+                />
               ) : null}
             </FlexGrid>
           );
@@ -65,12 +68,13 @@ const CommentList = (props) => {
 };
 const FlexGrid = styled.div`
   display: flex;
-  max-width: 750px;
+  max-width: ${(props) => (props.maxWidth ? props.maxWidth : "750px")};
+  ${(props) => (props.minWidth ? `min-width:${props.minWidth};` : null)};
   ${(props) => (props.color ? `background-color:${props.color};` : null)};
   ${(props) => (props.margin ? `margin:${props.margin};` : null)};
   ${(props) => (props.direction ? `flex-direction:${props.direction};` : null)};
-  ${(props) => (props.align ? `align-items:${props.align};` : null)}
-  ${(props) => (props.justify ? `justify-content:${props.justify};` : null)}
-  ${(props) => (props.padding ? `padding:${props.padding};` : null)}
+  ${(props) => (props.align ? `align-items:${props.align};` : null)};
+  ${(props) => (props.justify ? `justify-content:${props.justify};` : null)};
+  ${(props) => (props.padding ? `padding:${props.padding};` : null)};
 `;
 export default CommentList;
